Parse.ReadZepto.skipRestObj line-mode optimization
redefined Parse.ReadZepto.skipRestObj as `pure ()` and added
inline pragma.  Under current driver architecture, the Zepto
version of the parser is used only in line-mode parsing, and
so the remainder of the input provided to Zepto to parse
is precisely the remainder of one object; skipping the rest
of an object is therefore as simple as ignoring the rest
of the input.

If Zepto-based parsing is ever used in block-mode or similar,
this optimization would be semantically unsound, and so the
original version of the skipRestObj is preserved (commented out)
under the new definition.