Gated channels for parallel streaming
Introduced paradigm of converting between streams and channels to allow
serial components of the pipeline with different flow-rates to run
in parallel rather than bottle-neck on the slowest component. Expected
to yield highest performance improvement for http+gzip, otherwise may
in fact be a pessimization for stdin+gzip or http+raw

Deprecated list-version of stream bundle, as well as --vector
command-line option, as the vector version is faster and cleaner.
Removed bloat associated with list-vector dichotomy and redefined
Bundle as a type alias for Vector (kept bundle functions)

Introduced ChanBounded data-type that holds a channel, a TVar
representing the number of queued (non-terminal) values, and
a pure integral upper bound on the buffered size

Reimplemented Driver to replace Chan+TVar couplings with ChanBounded

Created separate constants for upper bounds on pre-processing 'gate' channels
and post-processing 'work' channels