Implements mapParses for custom post-processing
Adds block-mode function `mapParses` that takes the same arguments
as `lineParseFold` (modulo different Parser) and produces the final
output in the IO monad. Unlike `lineParseFold`, this does not involve
a stream as work is not chunked, and so the result will tend to be
far 'larger' for the same input.