Reimplements Global/Internal, adds Zepto IO
Continued updates of haddock documentation of top-level modules

Implements a record type GlobalConf for size of batches, number of
workers, and BoundedChan sizes. Reimplements code that formerly depended
on previous implementation of constant values to use either a ReaderT
of a GlobalConf, or default to the 'defaultGlobalConf' values. This
methodology should at some point be improved.

Implements 'lineParseFoldIO', which takes an `IO b` as the return-type
of the finalizer function, but in other respects is an identical
interface to 'lineParseFold'

Exports top-level parser-combinators strTo{Zepto,Atto,Atto'} that take
a query key as a string and perform conversion to [ParseClass]
internally, obviating the need for exporting qkey and requring users
to compute the [ParseClass] conversion to pass into a parser-combinator
(i.e. seekInObj and variants)

TODO: consider merging Driver into Zepto